Do not write source directory changes
Currently we store these json files compressed. we do this on every
update and insert. Since I started doing this I have added the html in
to the index. Effectively duplicating the jsonc files sans the html to
text. I updated the ..update method to account for this by reparsing the
stored html. This will allow us to also progressively apply improvements
to the text parsing.

When I am forced to change the index the user will have to run to
commands.

personal_search --backup_index
personal_search --import_index

Only the current build of the main bin will need both the new and old
index. How can i keep track of the "versions"...